+++ src/App.tsx
@@ -1,21 +1,23 @@
 import { createHashRouter, RouterProvider } from 'react-router-dom';
 import { SWRConfig } from 'swr';
 import Console from './app/console/Console';
+import ErrorLayout from './app/ErrorLayout';
 import EditFlag from './app/flags/EditFlag';
 import Evaluation from './app/flags/Evaluation';
 import Flag, { flagLoader } from './app/flags/Flag';
 import Flags from './app/flags/Flags';
 import NewFlag from './app/flags/NewFlag';
-import NotFound from './app/NotFound';
+import Layout from './app/Layout';
+import NotFoundLayout from './app/NotFoundLayout';
 import NewSegment from './app/segments/NewSegment';
 import Segment, { segmentLoader } from './app/segments/Segment';
 import Segments from './app/segments/Segments';
-import Layout from './components/Layout';
 
 const router = createHashRouter([
   {
     path: '/',
     element: <Layout />,
+    errorElement: <ErrorLayout />,
     children: [
       {
         element: <Flags />,
@@ -65,7 +67,7 @@ const router = createHashRouter([
   },
   {
     path: '*',
-    element: <NotFound />
+    element: <NotFoundLayout />
   }
 ]);
