+++ src/app/Layout.tsx
@@ -1,10 +1,10 @@
 import { useState } from 'react';
 import { Outlet } from 'react-router-dom';
-import ErrorNotification from './ErrorNotification';
-import { ErrorProvider } from './ErrorProvider';
-import Footer from './Footer';
-import Header from './Header';
-import Sidebar from './Sidebar';
+import ErrorNotification from '../components/ErrorNotification';
+import { ErrorProvider } from '../components/ErrorProvider';
+import Footer from '../components/Footer';
+import Header from '../components/Header';
+import Sidebar from '../components/Sidebar';
 
 // const userNavigation = [{ name: "Sign out", href: "#" }];
